pyproject.toml
[build-system]
requires = ["hatchling>=1.21.1"]
build-backend = "hatchling.build"

[project]
name = "oauth-tester"
version = "0.1.0"
description = "Local HTTPS OAuth/OIDC tester with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Team" }]
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.1",
  "authlib==1.3.1",
  "jinja2==3.1.4",
  "python-dotenv==1.0.1",
  "pydantic-settings==2.3.4",
  "httpx==0.27.0",
  "itsdangerous==2.2.0",
  "starlette==0.37.2",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.2",
  "pytest-asyncio==0.23.8",
  "respx==0.21.1",
]

[project.scripts]
oauth-tester = "oauth_tester.main:main"

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/oauth_tester"]

[tool.hatch.build.targets.sdist]
include = ["src/oauth_tester", "README.md", "pyproject.toml", "Makefile", "Dockerfile"]
